Add Windows 7 support
.net 4.5 is installed as a prerequisite for boxstarter, and Powershell
v4 is installed to more closely align the cmdlets available in Server 2012

+++ scripts/package.ps1
@@ -1,5 +1,11 @@
+$OS = Get-WmiObject -Class win32_OperatingSystem -namespace "root\CIMV2"
+
 Enable-RemoteDesktop
-Set-NetFirewallRule -Name RemoteDesktop-UserMode-In-TCP -Enabled True
+if ($OS.Version -eq "6.1.7601") {
+    C:\Windows\System32\netsh.exe advfirewall firewall add rule name="Remote Desktop" dir=in localport=3389 protocol=TCP action=allow
+    } else {
+    Set-NetFirewallRule -Name RemoteDesktop-UserMode-In-TCP -Enabled True
+}
 
 Write-BoxstarterMessage "Removing page file"
 $pageFileMemoryKey = "HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\Memory Management"
@@ -7,13 +13,24 @@ Set-ItemProperty -Path $pageFileMemoryKey -Name PagingFiles -Value ""
 
 Update-ExecutionPolicy -Policy Unrestricted
 
-Write-BoxstarterMessage "Removing unused features..."
-Remove-WindowsFeature -Name 'Powershell-ISE'
-Get-WindowsFeature | 
-? { $_.InstallState -eq 'Available' } | 
-Uninstall-WindowsFeature -Remove
+# cmdlet not available in win 7
+if ($OS.Version -ne "6.1.7601") {
+    Write-BoxstarterMessage "Removing unused features..."
+    Remove-WindowsFeature -Name 'Powershell-ISE'
+    Get-WindowsFeature | 
+    ? { $_.InstallState -eq 'Available' } | 
+    Uninstall-WindowsFeature -Remove
+}
 
-Install-WindowsUpdate -AcceptEula
+if ($OS.Version -eq "6.1.7601") {
+    # Skipping updates specific to Win10 upgrade & telemetry noted at:
+    # https://github.com/bmrf/tron/blob/d7c8e00a4300bcffa6095aa9361850f7430fd3d2/tron.bat#L1301-L1326
+    # (KB971033/KB2952664/KB3021917/KB3068708/KB3075249/KB3080149)
+    # This should prevent the 3+GiB KB3035583 from being downloaded
+    Install-WindowsUpdate -criteria "IsHidden=0 and IsInstalled=0 and Type='Software' and BrowseOnly=0 and UpdateID!='62c68477-0f45-46aa-8af8-d0c189d6dd8e' and UpdateID!='761d6bfd-33b5-4988-9d0c-417f0284168f' and UpdateID!='89febea4-e23a-4ff5-9ca9-d4fc9e768a70' and UpdateID!='0cd9efd9-d371-4e7d-8381-15ae5b55ea79' and UpdateID!='6cc5cc49-03a2-4609-882e-7889c547814e' and UpdateID!='48aa5065-93fc-4cc8-b071-80cb1da35f7b'" -AcceptEula
+    } else {
+    Install-WindowsUpdate -AcceptEula
+}
 if(Test-PendingReboot){ Invoke-Reboot }
 
 Write-BoxstarterMessage "Cleaning SxS..."
@@ -36,7 +53,12 @@ Dism.exe /online /Cleanup-Image /StartComponentCleanup /ResetBase
     }
 
 Write-BoxstarterMessage "defragging..."
-Optimize-Volume -DriveLetter C
+if ($OS.Version -eq "6.1.7601") {
+    Defrag.exe c: /H
+    } else {
+    Optimize-Volume -DriveLetter C
+}
+
 
 Write-BoxstarterMessage "0ing out empty space..."
 wget http://download.sysinternals.com/files/SDelete.zip -OutFile sdelete.zip
@@ -45,7 +67,11 @@ wget http://download.sysinternals.com/files/SDelete.zip -OutFile sdelete.zip
 ./sdelete.exe /accepteula -z c:
 
 mkdir C:\Windows\Panther\Unattend
-copy-item a:\postunattend.xml C:\Windows\Panther\Unattend\unattend.xml
+if ($OS.Version -eq "6.1.7601") {
+    copy-item a:\postunattendwin7.xml C:\Windows\Panther\Unattend\unattend.xml
+    } else {
+    copy-item a:\postunattend.xml C:\Windows\Panther\Unattend\unattend.xml
+}
 
 Write-BoxstarterMessage "Recreate pagefile after sysprep"
 $System = GWMI Win32_ComputerSystem -EnableAllPrivileges
@@ -53,10 +79,20 @@ $System.AutomaticManagedPagefile = $true
 $System.Put()
 
 Write-BoxstarterMessage "Setting up winrm"
-Set-NetFirewallRule -Name WINRM-HTTP-In-TCP-PUBLIC -RemoteAddress Any
-Enable-WSManCredSSP -Force -Role Server
+if ($OS.Version -eq "6.1.7601") {
+    C:\Windows\System32\netsh.exe advfirewall firewall add rule name="WinRM-HTTP" dir=in localport=5985 protocol=TCP action=allow
+    } else {
+    Set-NetFirewallRule -Name WINRM-HTTP-In-TCP-PUBLIC -RemoteAddress Any
+}
 
-Enable-PSRemoting -Force -SkipNetworkProfileCheck
+if ($OS.Version -eq "6.1.7601") {
+    Enable-PSRemoting -Force -SkipNetworkProfileCheck
+    Enable-WSManCredSSP -Force -Role Server
+    } else {
+    Enable-WSManCredSSP -Force -Role Server
+
+    Enable-PSRemoting -Force -SkipNetworkProfileCheck
+}
 winrm set winrm/config/client/auth '@{Basic="true"}'
 winrm set winrm/config/service/auth '@{Basic="true"}'
 winrm set winrm/config/service '@{AllowUnencrypted="true"}'
